Move test files next to their respective source
